Allow file config independent rebuilds
Adds isPrimary flag to the file config. When true this will rebuild all
configs for a given mode. When false will ignore changes to the inputs
when undefined will continue the default behavior of separate rebuilds.

This is intended for plugins that need to reduce the amount of
duplicated work from different configs, i.e. the stylus plugin.